# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SDBA Admin System API
  description: |
    API documentation for the SDBA Admin System.
    
    ## Authentication
    Admin endpoints require authentication via Supabase Auth. Include the session cookie in requests.
    
    ## CSRF Protection
    All state-changing requests (POST, PUT, DELETE, PATCH) require a CSRF token in the `X-CSRF-Token` header.
    Get the CSRF token from the `/api/csrf-token` endpoint or from the `__Host-csrf-token` cookie.
    
    ## Rate Limiting
    - Public API: 10 requests per 10 seconds per IP
    - Admin API: 100 requests per minute per authenticated user
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "ok": false,
      "error": "Error message",
      "code": "ERROR_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: SDBA Admin Support
    email: support@sdba.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/admin/list:
    get:
      tags:
        - Admin
      summary: List registrations
      description: Get a paginated list of registrations with filtering and search. Requires admin authentication.
      operationId: listRegistrations
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page (max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: q
          in: query
          description: Search term (searches team_name, org_name, email_1, team_code)
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [all, pending, approved, rejected]
            default: all
        - name: event
          in: query
          description: Filter by event type
          required: false
          schema:
            type: string
            enum: [all, tn, wu, sc]
            default: all
        - name: season
          in: query
          description: Filter by season (year)
          required: false
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
      responses:
        '200':
          description: List of registrations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
              example:
                ok: true
                page: 1
                pageSize: 50
                total: 100
                items:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    season: 2025
                    event_type: "tn"
                    division_code: "M"
                    category: "men_open"
                    option_choice: "Option 1"
                    team_code: "S25-M001"
                    team_name: "Team Alpha"
                    org_name: "Example Org"
                    status: "pending"
                    created_at: "2025-01-01T00:00:00Z"
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie (set automatically on login)
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token (get from /api/csrf-token or __Host-csrf-token cookie)
  schemas:
...
    ListResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        page:
          type: integer
          description: Current page number
        pageSize:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        items:
          type: array
          items:
            $ref: '#/components/schemas/RegistrationItem'
    RegistrationItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        season:
          type: integer
        event_type:
          type: string
          enum: [tn, wu, sc]
        division_code:
          type: string
        category:
          type: string
        option_choice:
          type: string
          enum: [Option 1, Option 2]
        team_code:
          type: string
        team_name:
          type: string
        org_name:
          type: string
        org_address:
          type: string
        manager_name:
          type: string
        manager_email:
          type: string
        manager_mobile:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        approved_by:
          type: string
          format: uuid
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        detail:
          type: array
          items:
            type: object
          description: Validation error details (only for 422 errors)
  responses:
...
    Forbidden:
      description: Forbidden - authentication required or insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error: "Forbidden"
            code: "FORBIDDEN"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"